+++ NEWS.md
@@ -1,3 +1,24 @@
+# parameters 0.6.1
+
+## General
+
+* Remove 'Zelig' from suggested packages, as it was removed from CRAN.
+
+## Changes to functions
+
+### model_parameters()
+
+* `model_parameters()` now also transforms standard errors when `exponentiate = TRUE`.
+* `model_parameters()` for `anova()` from mixed models can now also compute effect sizes like eta squared.
+* `model_parameters()` for `aov()` gains a `type`-argument to compute type-1, type-2 or type-3 sums of squares.
+* `model_parameters()` for Bayesian models gains a `standardize` argument, to return standardized parameters from the posterior distribution.
+* Improved `print()` method for `model_parameters()` for nested `aov()` (repeated measurements).
+* You can now control whether `demean()` should add attributes to indicate within- and between-effects. This is only relevant for the `print()`-method of `model_parameters()`.
+
+## Bug fixes
+
+* Fixed `model_parameters()` for `anova()` from *lmerTest* models.
+
 # parameters 0.6.0
 
 ## Breaking changes